refactor(chat): decompose controller into functional orchestrators (#459)
